+++ Documentation/Changelog.md
@@ -13,6 +13,7 @@ To upgrade from TDW v1.11 to v1.12, read [this guide](upgrade_guides/v1.11_to_v1
 ### Docker
 
 - Updated Docker container to Ubuntu 22.04. Added instructions to `pip install tdw`.
+- Added: `Docker/build.sh`
 
 ## v1.12.9